LPS-68243 it should be int not long

+++ portal-impl/src/com/liferay/portal/util/PropsValues.java
@@ -636,7 +636,7 @@ public class PropsValues {
 
 	public static final String INDEX_DATE_FORMAT_PATTERN = PropsUtil.get(PropsKeys.INDEX_DATE_FORMAT_PATTERN);
 
-	public static final long INDEX_DEFAULT_INTERVAL = GetterUtil.getLong(PropsUtil.get(PropsKeys.INDEX_DEFAULT_INTERVAL));
+	public static final int INDEX_DEFAULT_INTERVAL = GetterUtil.getInteger(PropsUtil.get(PropsKeys.INDEX_DEFAULT_INTERVAL));
 
 	public static final boolean INDEX_DUMP_COMPRESSION_ENABLED = GetterUtil.getBoolean(PropsUtil.get(PropsKeys.INDEX_DUMP_COMPRESSION_ENABLED));
